dynamic top4 ranking
