+++ frontend/src/components/Top4Crypto/data/data.json
@@ -1,0 +1,30 @@
+[
+  {
+    "rank": 1,
+    "name": "Bitcoin",
+    "price": 7340.2,
+    "currency": "euro",
+    "change": "+0.07%"
+  },
+  {
+    "rank": 2,
+    "name": "Etherum",
+    "price": 156.35,
+    "currency": "euro",
+    "change": "+0.12%"
+  },
+  {
+    "rank": 3,
+    "name": "Bitcoin Cash",
+    "price": 308.7,
+    "currency": "euro",
+    "change": "+0.86%"
+  },
+  {
+    "rank": 4,
+    "name": "Litecoin",
+    "price": 51.77,
+    "currency": "euro",
+    "change": "-0.19%"
+  }
+]
